Move layers defined in contrib/classification/ofa_xception/modules.py to contrib/common/ofa/layers.py
